Refactor: 플레이리스트 및 비디오 DB 저장 로직 service -> api 폴더로 이동, 사용자 정보 컴포넌트에서 조회 후 전달
